src/requirements.txt
discord.py
giphy_client
random